requirements.txt
arrow==1.2.3
asgiref==3.4.1
async-timeout==4.0.2
blessed==1.20.0
certifi==2023.7.22
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.22
django-filter==21.1
django-import-export==2.7.1
django-otp==1.1.6
django-picklefield==3.1
django-q==1.3.9
django-redis==5.4.0
django-simpleui==2023.8.28
djangorestframework==3.14.0
et-xmlfile==1.1.0
idna==3.4
importlib-metadata==4.8.3
itypes==1.2.0
Jinja2==3.0.3
Markdown==3.3.7
MarkupPy==1.14
MarkupSafe==2.0.1
mysqlclient==2.1.1
odfpy==1.4.1
openpyxl==3.1.2
packaging==21.3
Pillow==8.4.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.3.1
redis==3.5.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.4
tablib==3.1.0
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.18
wcwidth==0.2.8
xlrd==2.0.1
xlwt==1.3.0
zipp==3.6.0
